# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API is a user-friendly solution designed to serve as the foundation for more complex APIs.",
        "title": "Go API",
        "contact": {
            "name": "Raul del Aguila",
            "url": "https://github.com/raulaguila",
            "email": "email@email.com"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/auth": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "User authenticated",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User authenticated",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "enum": [
                            "en-US",
                            "pt-BR"
                        ],
                        "type": "string",
                        "default": "en-US",
                        "description": "Request language",
                        "name": "Accept-Language",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_core_dto.UserOutput"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "User refresh",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User refresh",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "enum": [
                            "en-US",
                            "pt-BR"
                        ],
                        "type": "string",
                        "default": "en-US",
                        "description": "Request language",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "type": "boolean",
                        "description": "Expire token",
                        "name": "expire",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_core_dto.AuthOutput"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "User authentication",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User authentication",
                "parameters": [
                    {
                        "enum": [
                            "en-US",
                            "pt-BR"
                        ],
                        "type": "string",
                        "default": "en-US",
                        "description": "Request language",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "description": "Credentials model",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_core_dto.LoginInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_core_dto.AuthOutput"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    }
                }
            }
        },
...
        "/role": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get roles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "Get roles",
                "parameters": [
                    {
                        "enum": [
                            true,
                            false
                        ],
                        "type": "boolean",
                        "default": true,
                        "description": "Skip auth",
                        "name": "X-Skip-Auth",
                        "in": "header"
                    },
                    {
                        "enum": [
                            "en-US",
                            "pt-BR"
                        ],
                        "type": "string",
                        "default": "en-US",
                        "description": "Request language",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "list_root",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "with_permissions",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_core_dto.PaginatedOutput-github_com_raulaguila_go-api_internal_core_dto_RoleOutput"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Insert role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "Insert role",
                "parameters": [
                    {
                        "enum": [
                            true,
                            false
                        ],
                        "type": "boolean",
                        "default": true,
                        "description": "Skip auth",
                        "name": "X-Skip-Auth",
                        "in": "header"
                    },
                    {
                        "enum": [
                            "en-US",
                            "pt-BR"
                        ],
                        "type": "string",
                        "default": "en-US",
                        "description": "Request language",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "description": "Role model",
                        "name": "role",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_core_dto.RoleInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_core_dto.RoleOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete roles by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "Delete roles by ID",
                "parameters": [
                    {
                        "enum": [
                            true,
                            false
                        ],
                        "type": "boolean",
                        "default": true,
                        "description": "Skip auth",
                        "name": "X-Skip-Auth",
                        "in": "header"
                    },
                    {
                        "enum": [
                            "en-US",
                            "pt-BR"
                        ],
                        "type": "string",
                        "default": "en-US",
                        "description": "Request language",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "description": "Roles ID",
                        "name": "ids",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_core_dto.IDsInput"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_raulaguila_go-api_internal_adapter_driver_rest_presenter.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                },
                "object": {}
            }
        },
        "github_com_raulaguila_go-api_internal_core_dto.AuthOutput": {
            "type": "object",
            "properties": {
                "accesstoken": {
                    "type": "string"
                },
                "refreshtoken": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_raulaguila_go-api_internal_core_dto.UserOutput"
                }
            }
        },
        "github_com_raulaguila_go-api_internal_core_dto.IDsInput": {
            "type": "object",
            "required": [
                "ids"
            ],
            "properties": {
                "ids": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "github_com_raulaguila_go-api_internal_core_dto.LoginInput": {
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "expiration": {
                    "type": "boolean"
                },
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "github_com_raulaguila_go-api_internal_core_dto.PaginatedOutput-github_com_raulaguila_go-api_internal_core_dto_RoleOutput": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_raulaguila_go-api_internal_core_dto.RoleOutput"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_raulaguila_go-api_internal_core_dto.PaginationOutput"
                }
            }
        },
...
        "github_com_raulaguila_go-api_internal_core_dto.PaginationOutput": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_raulaguila_go-api_internal_core_dto.RoleInput": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 4
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_raulaguila_go-api_internal_core_dto.RoleOutput": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "github_com_raulaguila_go-api_internal_core_dto.UserOutput": {
            "type": "object",
            "properties": {
                "corp_id": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "new": {
                    "type": "boolean"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_raulaguila_go-api_internal_core_dto.RoleOutput"
                    }
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and the JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}